modularise backbone following browserify's slightly strange stipulations RE requiring backbone per every backbone module
